requirements.txt
altgraph==0.17.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-bootstrap-templates==1.2.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
dash_uploader==0.6.1
firebird-base==1.8.0
firebird-driver==1.10.6
Flask==3.0.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==11.0.0
plotly==5.24.1
protobuf==5.28.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
retrying==1.3.4
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.36
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
unicode==2.9
Unidecode==1.3.8
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
